add mode to kDataFrameMQF constructor (swig)
